+++ relay-ourlogs/src/lib.rs
@@ -1,4 +1,4 @@
-//! Structs and functions needed to ingest OpenTelemetry logs.
+//! Structs and functions needed to ingest Sentry logs.
 
 #![warn(missing_docs)]
 #![doc(
@@ -6,8 +6,10 @@
     html_favicon_url = "https://raw.githubusercontent.com/getsentry/relay/master/artwork/relay-icon.png"
 )]
 
-pub use crate::ourlog::otel_to_sentry_log;
-pub use crate::ourlog::ourlog_merge_otel;
-pub use opentelemetry_proto::tonic::logs::v1::LogRecord as OtelLog;
-
 mod ourlog;
+mod size;
+
+pub use self::ourlog::otel_to_sentry_log;
+pub use self::ourlog::ourlog_merge_otel;
+pub use self::size::calculate_size;
+pub use opentelemetry_proto::tonic::logs::v1::LogRecord as OtelLog;